fix(build): Benchmark compilation (#14552)
Summary:
With the C++20 change using volatile is an error.
There was one benchmark that did use this. but it didn’t come up in the CI which didn’t seem to build the benchmarks at all. As a result the CI needed a fix to ensure proper compilation.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14552

Reviewed By: bikramSingh91

Differential Revision: D82456701

Pulled By: kgpai

fbshipit-source-id: e3c68d3c8be2aaf200390cba993a3ebadb4d3408