refactor(cudf): Add CudfHiveConnector (#14294)
Summary:
Add CudfHiveConnector, which inherits from the HiveConnector.
We want to use the HiveConnector implementation as a base for the CudfConnector since it is well established, and it gives us the flexibility to fallback to CPU when needed.

This PR also disables cuDF based parquet writer for now to limit the scope of changes.

Fixes: https://github.com/facebookincubator/velox/issues/14183

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14294

Reviewed By: gggrace14

Differential Revision: D82649332

Pulled By: xiaoxmeng

fbshipit-source-id: 5c230db9fce729eeec67af1d2c0dac8bdb66b9ac