fix(parquet): Read unannotated array (#13864)
Summary:
When the element type is scalar type, a check in `convertType` requires the
requested type must not be array. However, an unannotated array in Parquet
is a repeated field that is not explicitly marked as a LIST logical type.
To enable reading of unannotated arrays, this PR verifies the compatibility of
their element types.

Follow-up for https://github.com/facebookincubator/velox/pull/13620.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/13864

Reviewed By: gggrace14

Differential Revision: D82649435

Pulled By: xiaoxmeng

fbshipit-source-id: 6dd331de48534c33127679a96331a6c88035ed2d