feat(abfs): Add DynamicSasTokenClientProvider to support dynamic SAS token renewal (#14045)
Summary:
In the Java implementation, user can implement and configure the
`SasTokenProvider` for SAS token generation with interface
https://github.com/apache/hadoop/blob/trunk/hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/extensions/SASTokenProvider.java.
Once the generated token has expired or is near expiry, it will be
renewed by acquiring a new token from the provider.

Currently in Velox, the SAS token is configured as a fixed token,
and expiration is not considered. This patch adds
`DynamicSasTokenClientProvider` to handle the SAS token renewal
logic by updating the reader/writer client with a renewed SAS and
adds `AbfsSasTokenProvider` interface to provide the SAS token.
User can choose to register the `DynamicSasTokenClientProvider`
with a custom `AbfsSasTokenProvider` implementation for
accounts that requires dynamic SAS token renewal.

relies on https://github.com/facebookincubator/velox/pull/14281

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14045

Reviewed By: mbasmanova

Differential Revision: D82029830

Pulled By: Yuhta

fbshipit-source-id: 20a6a376324303bbac3c04c904a8d53f606555fe