refactor: Do not hold shared_ptr of the TaskQueue (#14896)
Summary:
Cursor owns TaskQueue and Task. Both of the Task’s Consumer lambda and
OnError, the lambda captures the shared_ptr of the TaskQueue. If the task fails
while there are leftover Vectors in the TaskQueue, it may even crash if leaf pools are
destroyed before the Vectors in the TaskQueue.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/14896

Reviewed By: kewang1024

Differential Revision: D82746614

Pulled By: xiaoxmeng

fbshipit-source-id: 1d66351d0529fe4aef60a0129e19e96bc6c3d32b