[native] Fix incorrect initialization of 'folly::Promise' object
Fixes https://github.com/prestodb/presto/issues/26094

For a `folly::Promise` object that is expected to be overwritten, we
should use `folly::makeEmpty()` to initialize it (it is 'invalid')
instead of the default constructor (it will be 'valid' but
'not fulfilled', assigning to it will cause an exception). Creating an
exception triggers a stack unwind, which can saturate the CPU in
high-concurrency scenarios, causing significant performance issues.