Rewrite Geometry to Varchar type for verifier temp tables (#26140)
Summary:

For queries being tested with verifier, if the output type is Geometry, the test case will be skipped due to a "NOT_SUPPORTED Type 'Geometry' cannot be written to a Hive table. Cast to a supported type such as VARCHAR or VARBINARY." error for both the control and test queries. This change rewrites the column type from Geometry to Varchar using the ST_AsText function, which is deterministic.

Differential Revision: D83139143