[presto][native] Init system used memory cache at the beginning of PeriodicMemoryChecker callback
Summary:
in Currently `PeriodicMemoryChecker` and descendants use these 2 methods to get current system used memory:
- `systemUsedMemoryBytes`:
```
int64_t systemUsedMemoryBytes() {
  const auto currentMemBytes = ...;
  cachedSystemUsedMemoryBytes_ = currentMemBytes;
  return currentMemBytes;
}
```
- `cachedSystemUsedMemoryBytes`
```
int64_t cachedSystemUsedMemoryBytes() const {
  return cachedSystemUsedMemoryBytes_;
}
```
Both methods are `protected` and don't become public in descendants, so their usage is limited by the classes itself. We get value from cache only if there is a prior `systemUsedMemoryBytes` call, before that the cache is empty. All `systemUsedMemoryBytes` and `cachedSystemUsedMemoryBytes` calls only happen inside the callback, this makes it easy to maintain the invariants.

### The problem
In one of the upcoming diffs I'll need an access to the value of current system used memory inside `periodicCb()` implementation. The straightforward approach would be to just call `systemUsedMemoryBytes`, but it's a syscall and I noticed, that we can do some optimization in that area. We can do `systemUsedMemoryBytes()` call as the very first line in the callback and then everywhere below use cached value or actual value depends on needs. As example in `PeriodicMemoryChecker::pushbackMemory` we can also get the value from cache, and then update it by calling `systemUsedMemoryBytes()` after the pushback, which we're already doing in the code.

Differential Revision: D83218739