feat: Add probe/build geometries and radius as members of SpatialJoinNode (#25998)
Summary:
Add three fields to SpatialJoinNode:
```
probeGeometryVariable: VariableReferenceExpression
buildGeometryVariable: VariableReferenceExpression
radiusVariable: Optional<VariableReferenceExpression>
```

These are VREs to the probe and build side geometries that the spatial join will need to make envelopes for the
spatial index.  The radius variable, if present, will increase the size of the build side envelope.

Currently, when Presto Java extracts a SpatialJoinNode, it has to check that there's a probe/build pair of geometries that it will take envelopes of and do a quick test of envelope intersection.  It then throws that information away, leaving only the join predicate `filter`. Then, LocalExecutionPlanner does similar logic again to extract these variables to actually build the spatial index and make the SpatialJoinOperator.

For velox, we would require implementing that logic a _third_ time, in particular doing several non-trivial checks to discover the geometries, and then failing in the unexpected case that this is not possible.  Instead, if we pull that logic up to ExtractSpatialJoin, _both_ the java LocalExecutionPlanner and the velox

No user-facing API changes.  It should not be a appreciable performance impact.

Test Plan:
No change of behavior expected.  Unit tests (`mvn test`).

- [x] Please make sure your submission complies with our [contributing guide](https://github.com/prestodb/presto/blob/master/CONTRIBUTING.md), in particular [code style](https://github.com/prestodb/presto/blob/master/CONTRIBUTING.md#code-style) and [commit standards](https://github.com/prestodb/presto/blob/master/CONTRIBUTING.md#commit-standards).
- [x] PR description addresses the issue accurately and concisely.  If the change is non-trivial, a GitHub Issue is referenced.
- [x] Documented new properties (with its default value), SQL syntax, functions, or other functionality.
- [x] If release notes are required, they follow the [release notes guidelines](https://github.com/prestodb/presto/wiki/Release-Notes-Guidelines).
- [x] Adequate tests were added if applicable.
- [ ] CI passed.

```
== NO RELEASE NOTE ==
```

Differential Revision: D82045314

Pulled By: jagill