refactor: Refactor ExtractSpatialJoin
In preparation for extracting geometry and radius variables, refactor the logic
in ExtractSpatialJoin. Specifically, merge newRadiusVariable and
newGeometryVariable methods (which did the same thing), add a more explicit
function to check which sides an expression's variables are on, and simplify
some logic with these.

This commit should contain no behavioral change, so that the next commit can
focus on exposing the new variables.