feat(geo): Add geometry and radius variables to SpatialJoinNode
In Java, the optimizer's ExtractSpatialJoins as well as the
LocalExecutionPlanner must find build and probe expressions that contain
geometries for the join to be planned or executed.  In ExtractSpatialJoins,
this work was then thrown away.  Velox also needs to do this when constructing
its SpatialJoinBuild/Probe.

Instead of adding planner logic in Velox, this extracts the already calculated
variables for probe and build geometries and optionally a radius and adds them
to SpatialJoinNode.  This will allow us to pass them to Velox.

Also update presto cpp protocol to reflect these new variables.