fix(ci): Free disk space action
The previous github action used did not work.
The reason is that it was run inside the
(dependency) container where the commands
did not run (sudo not found) as well as the
files to be deleted were not present.

This fix mounts the host path that contains the
files to be deleted into the container and
we can run custom commands to clean up disk
space.

The free-disk-space action uses sudo to remove packages
that are not needed. However, sudo is not found and as a result
does not clean up any disk space