feat(geo): Change Coordinator Geometry conversion to use JTS
When a Geometry object reaches the coordinator to be returned to a client, it must be JSON-serialization. We've chosen this representation to be WKT, and previously deserialized it with Esri.  As we move to GEOS-based serialization in Prestissimo, which is compatible with JTS deserialization in Java.  Change `GeometryType.getObjectValue` to use JTS instead, which also changes how it handles exceptions.

Also apply the same change to SphericalGeography.