feat(reader): Extend session property to control row size tracking for metalake (#26271)
Summary:


Due to extra integration uncertainty in the metalake path and some previous QB signals, we decide to separately control row size tracking for metalake with the session property.

Instead of adding an additional session property, we decide to extend the current one. However, due to changing the session property type breaking backward compatibility, we still ended up introducing a new query config. We will delete the deprecated bool session property as the 3rd diff in the stack. Since we don't yet have any real mitigation with the deprecated bool session property, we directly replace it with the new one.

Differential Revision: D84244621