fix: Avoid re-creating future in ScaledWriterScheduler
Do not re-create future when previous future is not done. This is needed
to avoid unnecessary scheduling overhead on the scheduled executor when
the query scheduler gets unblocked more frequently.

For example consider a simple query with two stages containing TableScan and
TableWrite. The TableScan scheduler may never block causing the ScaledWriterScheduler
to be unblocked before the underlying future is unblocked. This will
cause accumulation of tasks on the ScheduledExecutorService.

This semantic is consistent with SourcePartitionedScheduler. The SourcePartitionedScheduler
caches future underneath and allows the future returned to be cancelled
by the SqlQueryScheduler.