fix(native): Init system used memory cache at the beginning of PeriodicMemoryChecker callback (#26347)
## Description
The initial idea behind this change was to preload SystemMemoryUsage at
the beginning of the `PeriodicMemoryChecker` callback. If the memory is
preloaded, in the subsequent `periodicCb()` and `pushbackMemory()`
implementation calls, we can decide whether to use the preloaded value
or refetch it from the system.

For our use case, we want to report the value obtained to the monitoring
system. This will help in understanding the exact memory value used in
the pushback mechanism.

This PR was imported from Meta Phabricator: D83218739
The discussion there led to a desire to change the API to make it
clearer for users.

So in this diff I'm changing the API from:
```c++
  int64_t cachedSystemUsedMemoryBytes() const {
    return cachedSystemUsedMemoryBytes_;
  }

  virtual int64_t systemUsedMemoryBytes() = 0;
```
to:
```c++
  int64_t systemUsedMemoryBytes(bool fetchFresh = false);

  virtual void loadSystemMemoryUsage() = 0;
```
With the new API, we can clearly specify whether we want to preload the
value, retrieve it from the cache, or fetch the current one.

The API change only affects `PeriodicMemoryChecker` descendants. If this
PR broke your build, follow this guide to fix the code:
1) Replace `systemUsedMemoryBytes()` calls with
`systemUsedMemoryBytes(/*fetchFresh=*/true)`
2) Replace `cachedSystemUsedMemoryBytes()` calls with
`systemUsedMemoryBytes()`
3) Put `systemUsedMemoryBytes()` impl to `loadSystemMemoryUsage()` impl

## Test Plan
This PR is more like a refactoring, but I tested it in prod-like
environment in Meta.

```
== NO RELEASE NOTE ==
```