fix(native): Generation of thrift protocol
The thrift generation was broken because a line
that was expected to be removed was not and caused
a thrift parse error.

That line is "drift.recursive_reference = true,” which
was not caught by the regex attempting to remove it.

This PR also adds the required thrift parser to the
python installation of the dependency to use it in the CI.