fix(native): Incorrect session property hidden attribute
A session property can be declared hidden.
Some session property additions set a default value
instead of the hidden property. The integers were
interpreted as true.

This change fixes the attribute to true or false.
It is set to false for documented session properties and
true for undocumented properties at the time of addition.