feat(native): Add ability to detach worker when overloaded for too long (#26544)
## Description
We noticed that some bugs can lead to worker use most of the memory and
not let it go for a long time.
This results in memory manager mem-killing most of the queries as it
always on the verge of global OOM.
This makes the whole cluster pretty much non-operational, even worse,
failing big share of queries.

We add the logic similar to stuck drivers - when we are continuously
overloaded for some long time (controlled by the new config parameter)
we detach the worker to keep the cluster otherwise healthy.
The default config threshold is zero, which effectively disables the
feature.

```
== NO RELEASE NOTE ==
```