feat!: Allow connectors to pass through richer representations of commit metadata for query event listeners (#26331)
Summary:
Currently, the `Input` and `Output` query metadata classes retain two
source of connector-specific information that can be useful for
reporting via an `EventListener`:
```
Optional<Object> connectorInfo;
String serializedCommitOutput;
```
* `connectorInfo` can be cast back to the correct type in an
`EventListener` implementation, allowing rich access to the underlying
data
* `serializedCommitOutput` however, is serialized in a given format by
the `ConnectorCommitHandle` implementation, which makes it difficult to
correctly represent the reporting requirements in an EventListener
(which may need correlation with data in the `connectorInfo` result).

For example, `HiveCommitHandle` retains the lastDataCommitTime for each
partition in a simple array associated with the table name, where the
partition names are retained in the `HiveInputInfo` instance carried
through in connectorInfo. For these times to be mapped back to
individual partitions, the entries must be in the exact same order as
the entries in HiveInputInfo.

This change simply replaces the `serializedCommitOutput` property with
an `Optional<Object>` instance, providing parity with the
`connectorInfo`, and allowing `EventListener` implementations to cast
the commit handle back to the correct type for richer access to the
underlying data.

Differential Revision: D84382446

## Release Notes
```
== RELEASE NOTES ==

SPI Changes
* Replaces the ``String serializedCommitOutput`` argument with ``Optional<Object> commitOutput`` in the ``com.facebook.presto.spi.eventlistener.QueryInputMetadata`` and ``com.facebook.presto.spi.eventlistener.QueryOutputMetadata`` constructors
* Adds ``getCommitOutputForRead()`` and ``getCommitOutputForWrite()`` methods to ``ConnectorCommitHandle``, and deprecates the existing ``getSerializedCommitOutputForRead()`` and ``getSerializedCommitOutputForWrite()`` methods
```