fix(analyzer): materialized view join filtering (#26630)
Summary:

When there is a join query that involves materialized view, if the WHERE predicates includes columns of the join table, it throws a semantic exception about column cannot be resolved.

Example query:
SELECT * FROM mv JOIN table WHERE table_col1>10

The reason is that When Presto tries to optimize partition filtering for materialized views, it analyzes the WHERE clause using only the scope of the materialized view table. This fails when the WHERE clause contains columns from joined tables that haven't been processed yet.

This PR fixes it by only filtering out columns that belong to the materialized view. Also it uses analyzeExpression instead of analyzeWhere to not record this, so that it won't miss the join table columns with complete scope.

Differential Revision: D87164277