fix: Flaky test FunctionMetadataTest.mod (#26441)
## Description
<!---Describe your changes in detail-->
The constraint arrays in function metadata are created from unordered
map, so the element order may be different depending on the
implementation of the map. To make sure that the test consistently
passes regardless of implementation, I added code to sort the element of
the constraint arrays before the check to ensure that the function
metadata is equal to the expected value.

## Motivation and Context
<!---Why is this change required? What problem does it solve?-->
<!---If it fixes an open issue, please link to the issue here.-->
Resolves: #26440

## Impact
<!---Describe any public API or user-facing feature change or any
performance impact-->
No impact on public facing APIs, changes are only in test code

## Test Plan
<!---Please fill in how you tested your change-->
The unit test now passes (previously it failed on my Mac)

```
== NO RELEASE NOTE ==
```

---------

Signed-off-by: Ruth Kurniawati <ruth.kurniawati@gmail.com>