fix(analyzer): Materialized view with logical view and cte (#26713)
Summary:
If a materialized view is a part of a logical view, the logical view's
where predicate is not pushed down to materialized view so that it
doesn't check the overlap correctly. It caused the comparison between
mv's data and ALL base table data instead of the ones specified in the
query.

This diff fixes it by storing the where predicate when processing a
logical view. So mv can combine the where predicate in logical view as
well when getting mv status. It also fixes the issue during with using
the logical view/mv in cte.

Differential Revision: D87928199