refactor(native): Modify `PlanConversionFailureInfo` to `NativeSidecarFailureInfo` (#26720)
## Description
Refactors `PlanConversionFailureInfo` to `NativeSidecarFailureInfo` in
order to externalize its uses beyond the `NativePlanChecker` (and for
use in the `NativeExpressionOptimizer`).
Also exposes the helper method that converts
`protocol::ExecutionFailureInfo` to `protocol::NativeSidecarFailureInfo`
for use in native expression optimization endpoint, `v1/expressions` (to
be added).

## Motivation and Context
Needed for https://github.com/prestodb/presto/pull/26475.

## Impact
N/A.

## Test Plan
Existing tests for `NativePlanChecker`.


```
== NO RELEASE NOTE ==
```