perf(native-pos): Broadcast join: Improve broadcast files read performance (#26732)
Differential Revision: D88238987

We make 3 main changes in this diff to improve performance of broadcast
join file reads in native pos stack

- The footer read call for broadcast file has been moved out of
constructor and into a standalone function which is called once by the
first caller (`request`, `requestDataSizes`). This solves the issue of
ExchangeClient::addRemoteTaskId calling requestDataSize on each source
sequentially in a loop, which is causing performance regression, when
the number of files / sources are large , say 4096

- Make the `requestDataSizes` implementation async so that the
`addRemoteTaskId` calling `request` does not result in blocking I/O
call. This is currently resulting in a performance regression

- Add 2 new RuntimeMetrics to Exchange operator -
`broadcastExchangeSource.openFileAndReadFooterTime` and
`broadcastExchangeSource.fileReadWallTime` which measures the time to
read footer and the time to read the pages in the broadcast file


```
== NO RELEASE NOTE ==
```