chore: Cleanup native tests (#26154)
## Description
Cleans up tests from `presto-native-tests`.

## Motivation and Context
Helper methods used to test custom functions are only used in
`TestCustomFunctions` and should not be a part of `NativeTestUtils`.
Builder pattern should be used to get the `queryRunner` in tests and
function `NativeTestUtils.createNativeQueryRunner` should not be
overloaded with different arguments.
Enables tests with time type in `AbstractTestAggregationsNative`, since
it is now supported in Velox.

## Impact
NA


```
== NO RELEASE NOTE ==
```