feat(native): Add endpoint for expression optimization in sidecar (#26475)
## Description
To support constant folding and consistent semantics between the Presto
Java coordinator and the Presto C++ worker, it is necessary to use
consistent expression evaluation. To support this, a native expression
evaluation endpoint, `v1/expressions`, has been added to the Presto C++
sidecar. This endpoint leverages the `ExprOptimizer` in Velox to
optimize Presto expressions.

The optimized Velox `core::TypedExpr` returned by Velox's
`ExprOptimizer` is converted to a Presto `protocol::RowExpression` in
the Presto native sidecar with helper class
`VeloxToPrestoExprConverter`. The end to end flow between the
coordinator and sidecar is as follows:
```
RowExpression == (PrestoToVeloxExpr) ==> TypedExpr == (Velox ExprOptimizer optimize() API) ==> optimized TypedExpr == (VeloxToPrestoExprConverter) ==> optimized RowExpression
```

When `OptimizerLevel` is `EVALUATED` and an error is encountered during
expression evaluation, it is converted to `NativeSidecarFailureInfo`
with presto protocol. The list of optimized `RowExpression`s and
`NativeSidecarFailureInfo`s is returned along with their corresponding
indices in the input expressions list. With the fuzzer testing (see test
plan), we expect this endpoint to be ready for production workloads.

## Motivation and Context
The `native-sidecar-plugin` will implement the `ExpressionOptimizer`
interface from Presto SPI to utilize the `v1/expressions` endpoint on
the sidecar for optimizing Presto expressions using the native
expression evaluation engine.

The primary goal is to achieve consistency between C++ and Java
semantics for expression optimization. With this change, C++ functions
including UDFs can be used for constant folding of expressions in the
Presto planner.
Please refer to
[RFC-0006](https://github.com/prestodb/rfcs/blob/main/RFC-0006-expression-eval.md).

## Test Plan
Tests have been added by abstracting testcases from
`TestRowExpressionInterpreter` to an interface
`AbstractTestExpressionInterpreter`. This test interface is implemented
in `TestNativeExpressionInterpreter` to test the `v1/expressions`
endpoint on the sidecar end to end.

This feature is still in Beta, and to support production workloads with
complete certainty, the Velox expression fuzzer will be extended to test
this endpoint with fuzzer generated expressions in a follow-up PR. This
should surface any remaining bugs, such as in `VeloxToPrestoExpr`.
Note: The Velox `ExprOptimizer` and logical rewrites are already tested
with fuzzer.

## Release Notes

```
Prestissimo (Native Execution) Changes
* Add http endpoint `v1/expressions` in sidecar for expression optimization. See :doc:`/presto_cpp/sidecar`.
```

---------

Co-authored-by: Timothy Meehan <tim.meehan@ibm.com>