feat(testing): Add nondeterministic verifier retry (#26696)
Summary:

We have a verifier usecase where partitions may still be landing when the verifier is running. If the partition becomes available in between when the control and test query runs, the test is marked nondeterministic but still a failure. We actually aren't interested in this case and don't want to spend manual effort resolving it, we would rather re-run the test immediately in hopes that the nondeterminism goes away (e.g. that partitions don't continue to land). Adding a flag to that effect

There is not currently a mechanism to have an inline retry, so in this diff I am adding an additional CLI option to resubmit queries if determinism analysis has proven them to be nondeterministic

e.g., if fresh data has arrived and altered the results after the control query has run.

Reviewed By: NikhilCollooru

Differential Revision: D87831484