[presto_cpp] use hashTable caching in Broadcast joins (#26806)
Summary:

Velox introduced ability to Cache Hash Tables in the HashBuild operator.
This is useful in Broadcast joins as we can built the HashTable once per worker
and re-use it for all the join tasks that land on that worker.

Velox PRs: https://github.com/facebookincubator/velox/pull/15754
and https://github.com/facebookincubator/velox/pull/15768

This diff enables setting the `useCachedHashTable=true` during
velox `HashBuild` node construction in the cases that it is a 
broadcast join

Reviewed By: xiaoxmeng

Differential Revision: D88900941